added book now option

+++ tour_plan/rest/serializers/booking.py
@@ -23,6 +23,8 @@ class BookingSerializer(serializers.ModelSerializer):
     email = serializers.CharField(write_only=True, required=False)
     country = serializers.CharField(write_only=True, required=False)
     phone = serializers.CharField(write_only=True, required=False)
+    book_now = serializers.BooleanField(write_only=True, required=False)
+    single_item = BookingItemSerializer(write_only=True, required=False)
 
     class Meta:
         model = Booking
@@ -36,6 +38,8 @@ class BookingSerializer(serializers.ModelSerializer):
     def create(self, validated_data):
         with transaction.atomic():
             user_type = "guest"
+            is_book_now = validated_data.pop("book_now", None)
+            single_item = validated_data.pop("single_item", None)
             if self.context["request"].user.is_authenticated:
                 user_type = "user"
 
@@ -45,7 +49,8 @@ class BookingSerializer(serializers.ModelSerializer):
                 booking = Booking.objects.create(
                     user=user, user_type=user_type, traveler_details=traveler_details
                 )
-                cartitems = CartItem.objects.filter(user=user)
+                if not is_book_now:
+                    cartitems = CartItem.objects.filter(user=user)
             else:
                 full_name = validated_data.pop("full_name")
                 email = validated_data.pop("email")
@@ -60,26 +65,37 @@ class BookingSerializer(serializers.ModelSerializer):
                     user_type="guest",
                     traveler_details=traveler_details,
                 )
-                cartitems = CartItem.objects.filter(id__in=cart_item_ids)
+                if not is_book_now:
+                    cartitems = CartItem.objects.filter(id__in=cart_item_ids)
 
-            total_price = decimal.Decimal(0)
-            order_items_to_create = []
-            for item in cartitems:
-                total_price += item.item_price
-                order_items_to_create.append(
-                    BookingItem(
-                        booking=booking,
-                        tour_plan=item.tour_plan,
-                        time_slot=item.time_slot,
-                        num_adults=item.num_adults,
-                        num_children=item.num_children,
-                        num_infants=item.num_infants,
-                        item_price=item.item_price,
+            if not is_book_now:
+                total_price = decimal.Decimal(0)
+                order_items_to_create = []
+                for item in cartitems:
+                    total_price += item.item_price
+                    order_items_to_create.append(
+                        BookingItem(
+                            booking=booking,
+                            tour_plan=item.tour_plan,
+                            time_slot=item.time_slot,
+                            num_adults=item.num_adults,
+                            num_children=item.num_children,
+                            num_infants=item.num_infants,
+                            num_youth=item.num_youth,
+                            num_student_eu=item.num_student_eu,
+                            item_price=item.item_price,
+                        )
                     )
-                )
 
-            BookingItem.objects.bulk_create(order_items_to_create)
-            cartitems.delete()
-            booking.total_price = total_price
-            booking.save()
+                BookingItem.objects.bulk_create(order_items_to_create)
+                cartitems.delete()
+                booking.total_price = total_price
+                booking.save()
+            else:
+                booking_item = BookingItem.objects.create(
+                    booking=booking, **single_item
+                )
+                booking.total_price = booking_item.item_price
+                booking.save()
+
             return booking
